requirements.txt
Django==2.0
Python==3.6.1
djangorestframework==3.73
django-cors-headers==2.10
PyMySQL==0.7.11
Bootsrap==3.3.7
jquery==1.12.4


